refactor(backend): split operations.rs into per-endpoint submodules
